arm64: use 64-bit frame offsets and support large frames
Frame offsets were u32 and were truncated into the 12-bit immediate fields of
the prologue's `sub sp` and of every frame load/store, so a frame larger than
4 KB (sp adjust) or 32 KB (a slot access) was silently miscompiled — and the
offset type itself could not even represent a frame above 4 GB.

Widen the frame-offset type to u64 throughout (FrameLayout, Value::Slot,
LoadStoreUImm/LoadStoreFpUImm, alloca, the prologue/epilogue) and route every
stack reference through helpers that fall back from the immediate form to a
register-materialized offset when it does not fit:
- load/store: form `base + offset` in the x16 scratch (extended-register add so
  sp is allowed), then access [x16].
- sp adjustment and slot-address formation: same register fallback.

Adds the extended-register add/sub instruction (UXTX #0), which unlike the
shifted-register form permits sp as rd/rn, with clang-verified encodings and
unit tests, plus a large_frame FileCheck test and a 48 KB-frame runtime check.

+++ compiler/rustc_codegen_arm64/src/context.rs
@@ -49,8 +49,9 @@ pub enum Value {
     /// The address of a named symbol plus a byte offset (functions, statics, string literals).
     Sym { sym: Sym, offset: i64, ty: Type },
     /// A scalar value (integer, float, or pointer) materialized at the given `sp`-relative frame
-    /// offset. This is how the baseline keeps every non-constant SSA value live.
-    Slot { off: u32, ty: Type },
+    /// offset. This is how the baseline keeps every non-constant SSA value live. The offset is a
+    /// `u64` because frames are not bounded to 4 GiB.
+    Slot { off: u64, ty: Type },
     /// An undefined/poison value of the given type.
     Undef { ty: Type },
 }
